Fix tutorial + amélioration
Fix des problèmes du tutoriel lié au changement de la génération de la bombe, si on appuies sur espace pendant que le texte s'affiche il s'affiche d'un seul coup, et fix du bug lors de l'explosion de la bombe ou il était impossible d'y toucher

+++ Assets/Scripts/Utils/AudioManager.cs
@@ -29,7 +29,7 @@ public class AudioManager : MonoBehaviour
     /// </summary>
     /// <param name="soundEffect">Le sound effect a jouer</param>
     /// <param name="volume">Le volume du sound effect</param>
-    public void PlaySoundEffect(SoundEffects soundEffect, int volume = 1)
+    public void PlaySoundEffect(SoundEffects soundEffect, float volume = 1)
     {
         audioSourceSFX.PlayOneShot(soundEffects[(int)soundEffect], volume);
     }
